chore: added interface and changed methods to getPage and getPerPage
